mcp-server/requirements.txt
azure-identity>=1.15.0
azure-monitor-query>=1.3.0
